requirements.txt
numpy==1.11.0
pandas==0.23.4
tensorflow-gpu>=1.10.0
